feat: refactor the pipeline
